chore(pnp):  device pnp samples (#677)
